requirements.txt
requests>=2.31.0
python-binance>=1.0.19
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

